move fields/vector.hpp core/vector.hpp
