+++ include/NeoN/core/vector.hpp
