+++ include/NeoN/fields/fieldTypeDefs.hpp
@@ -2,9 +2,9 @@
 // SPDX-FileCopyrightText: 2023-2024 NeoN authors
 #pragma once
 
-#include "NeoN/fields/vector.hpp"
 #include "NeoN/core/primitives/scalar.hpp"
 #include "NeoN/core/primitives/vec3.hpp"
+#include "NeoN/core/vector.hpp"
 
 
 namespace NeoN
