+++ include/NeoN/finiteVolume/cellCentred/operators/surfaceIntegrate.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "NeoN/fields/vector.hpp"
+#include "NeoN/core/vector.hpp"
 #include "NeoN/core/executor/executor.hpp"
 #include "NeoN/core/input.hpp"
 #include "NeoN/dsl/operator.hpp"
